Use `curl --fail` to fetch sets and config
Currently, when fetching sets, if an invalid version is fed in, or an invalid
mirror, `curl` will still create the output sets or report that fetching
`install.conf` was successful.

With `--fail`, `curl` will now exit with failure on 404 and similar errors,
rather than creating sets filled with html 404 pages.

In the future, if error-checking is added, this makes it easier to ensure that
the script bails on a more complete set of `curl` failures.