Define `OPENBSD_TRUSTED_MIRROR` and `OPENBSD_MIRROR` later
Currently, if a user sets a different version of OpenBSD to fetch, then this
script will fetch 7.5 anyway, since the mirror URLs are defined up top, before
`OPENBSD_VERSION` is set  to the value that the user requested.

This diff sets the base of these URLs up top, then defines the full URL once we
know what version the user is requesting.